# backend-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VK Bot API",
        "description": "API для управления VK ботом",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/account/data/{id}": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Получить данные аккаунта",
                "description": "Получает расширенные данные пользователя ВКонтакте по его ID",
                "operationId": "fc4f3c393519929b642251020c9961fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя ВКонтакте",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "123456789"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные пользователя успешно получены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VkUserDataResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServerErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/account/newsfeed": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Получить новостную ленту аккаунта",
                "description": "Получает новостную ленту аккаунта ВКонтакте",
                "operationId": "65fbab6bfc9ef8035a3051f627d45e9e",
                "requestBody": {
                    "description": "Параметры для получения новостной ленты",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "account_id"
                                ],
                                "properties": {
                                    "account_id": {
                                        "description": "ID аккаунта",
                                        "type": "integer",
                                        "example": 123456789
                                    },
                                    "start_from": {
                                        "description": "Маркер для пагинации (не обязательный)",
                                        "type": "string",
                                        "example": "news_feed~3AA2ASgCBQPS8qggoQSUAdLyq-223862623_21815:1632147324:40",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Новостная лента успешно получена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsfeedResponseSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServerErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BaseSuccessResponse": {
                "description": "Базовая схема успешного ответа",
                "properties": {
                    "success": {
                        "description": "Статус успешности запроса",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "description": "Сообщение о результате операции",
                        "type": "string",
                        "example": "Операция выполнена успешно"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "description": "Схема ответа с ошибкой",
                "properties": {
                    "success": {
                        "description": "Статус успешности запроса",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Сообщение об ошибке",
                        "type": "string",
                        "example": "Произошла ошибка при выполнении операции"
                    }
                },
                "type": "object"
            },
            "ServerErrorResponse": {
                "description": "Схема ответа при внутренней ошибке сервера (500)",
                "properties": {
                    "message": {
                        "description": "Сообщение об ошибке",
                        "type": "string",
                        "example": "Внутренняя ошибка сервера при обработке запроса"
                    },
                    "exception": {
                        "description": "Класс исключения",
                        "type": "string",
                        "example": "App\\Exceptions\\ServiceException"
                    },
                    "file": {
                        "description": "Файл, в котором произошла ошибка",
                        "type": "string",
                        "example": "/var/www/html/vendor/package/src/Client.php"
                    },
                    "line": {
                        "description": "Строка, на которой произошла ошибка",
                        "type": "integer",
                        "example": 217
                    },
                    "trace": {
                        "description": "Трассировка стека вызовов",
                        "type": "array",
                        "items": {
                            "properties": {
                                "file": {
                                    "description": "Файл в трассировке",
                                    "type": "string",
                                    "example": "/var/www/html/vendor/package/src/Client.php"
                                },
                                "line": {
                                    "description": "Строка в трассировке",
                                    "type": "integer",
                                    "example": 185
                                },
                                "function": {
                                    "description": "Функция в трассировке",
                                    "type": "string",
                                    "example": "toException"
                                },
                                "class": {
                                    "description": "Класс в трассировке",
                                    "type": "string",
                                    "example": "ATehnix\\VkClient\\Client"
                                },
                                "type": {
                                    "description": "Тип вызова",
                                    "type": "string",
                                    "example": "::"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "NewsfeedResponseSchema": {
                "description": "Схема ответа для новостной ленты аккаунта",
                "properties": {
                    "success": {
                        "description": "Статус успешности запроса",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "description": "Сообщение о результате операции",
                        "type": "string",
                        "example": "Получены новые данные ленты"
                    },
                    "data": {
                        "description": "Данные новостной ленты",
                        "properties": {
                            "items": {
                                "description": "Список элементов новостной ленты",
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "type": {
                                            "description": "Тип элемента новостной ленты",
                                            "type": "string",
                                            "example": "post"
                                        },
                                        "source_id": {
                                            "description": "ID источника публикации",
                                            "type": "integer",
                                            "example": 3757268
                                        },
                                        "date": {
                                            "description": "Дата публикации в формате timestamp",
                                            "type": "integer",
                                            "example": 1754168183
                                        },
                                        "post_id": {
                                            "description": "ID поста",
                                            "type": "integer",
                                            "example": 4401
                                        },
                                        "owner_id": {
                                            "description": "ID владельца поста",
                                            "type": "integer",
                                            "example": 3757268
                                        },
                                        "text": {
                                            "description": "Текст поста",
                                            "type": "string",
                                            "example": ""
                                        },
                                        "attachments": {
                                            "description": "Прикрепления к посту",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "likes": {
                                            "description": "Информация о лайках",
                                            "properties": {
                                                "can_like": {
                                                    "description": "Можно ли поставить лайк",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "count": {
                                                    "description": "Количество лайков",
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "user_likes": {
                                                    "description": "Поставил ли пользователь лайк",
                                                    "type": "integer",
                                                    "example": 0
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "reposts": {
                                            "description": "Информация о репостах",
                                            "properties": {
                                                "count": {
                                                    "description": "Количество репостов",
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "user_reposted": {
                                                    "description": "Сделал ли пользователь репост",
                                                    "type": "integer",
                                                    "example": 0
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "comments": {
                                            "description": "Информация о комментариях",
                                            "properties": {
                                                "can_post": {
                                                    "description": "Можно ли комментировать",
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "count": {
                                                    "description": "Количество комментариев",
                                                    "type": "integer",
                                                    "example": 0
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "profiles": {
                                "description": "Список профилей пользователей",
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "id": {
                                            "description": "ID пользователя",
                                            "type": "integer",
                                            "example": 3757268
                                        },
                                        "first_name": {
                                            "description": "Имя пользователя",
                                            "type": "string",
                                            "example": "Анастасия"
                                        },
                                        "last_name": {
                                            "description": "Фамилия пользователя",
                                            "type": "string",
                                            "example": "Смирнова"
                                        },
                                        "screen_name": {
                                            "description": "Короткое имя пользователя",
                                            "type": "string",
                                            "example": "nastenchka"
                                        },
                                        "photo_50": {
                                            "description": "URL аватара 50x50",
                                            "type": "string",
                                            "format": "url"
                                        },
                                        "photo_100": {
                                            "description": "URL аватара 100x100",
                                            "type": "string",
                                            "format": "url"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "groups": {
                                "description": "Список групп",
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "id": {
                                            "description": "ID группы",
                                            "type": "integer",
                                            "example": 23064236
                                        },
                                        "name": {
                                            "description": "Название группы",
                                            "type": "string",
                                            "example": "Четкие Приколы"
                                        },
                                        "screen_name": {
                                            "description": "Короткое имя группы",
                                            "type": "string",
                                            "example": "ilikes"
                                        },
                                        "photo_50": {
                                            "description": "URL аватара группы 50x50",
                                            "type": "string",
                                            "format": "url"
                                        },
                                        "photo_100": {
                                            "description": "URL аватара группы 100x100",
                                            "type": "string",
                                            "format": "url"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "next_from": {
                                "description": "Маркер для получения следующей страницы",
                                "type": "string",
                                "example": "news_feed~3AA2ASgCBQPS8qgg5_-223862623_21815:1632147324:40"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "VkUserData": {
                "description": "Данные пользователя ВКонтакте с расширенной информацией",
                "properties": {
                    "id": {
                        "description": "ID пользователя",
                        "type": "integer",
                        "example": 123456789
                    },
                    "first_name": {
                        "description": "Имя пользователя",
                        "type": "string",
                        "example": "Иван"
                    },
                    "last_name": {
                        "description": "Фамилия пользователя",
                        "type": "string",
                        "example": "Иванов"
                    },
                    "screen_name": {
                        "description": "Короткое имя (никнейм) пользователя",
                        "type": "string",
                        "example": "ivanov",
                        "nullable": true
                    },
                    "photo_50": {
                        "description": "URL аватара 50x50",
                        "type": "string",
                        "example": "https://sun9-12.userapi.com/...",
                        "nullable": true
                    },
                    "photo_100": {
                        "description": "URL аватара 100x100",
                        "type": "string",
                        "example": "https://sun9-12.userapi.com/...",
                        "nullable": true
                    },
                    "photo_200": {
                        "description": "URL аватара 200x200",
                        "type": "string",
                        "example": "https://sun9-12.userapi.com/...",
                        "nullable": true
                    },
                    "status": {
                        "description": "Статус пользователя",
                        "type": "string",
                        "example": "Хорошее настроение!",
                        "nullable": true
                    },
                    "last_seen": {
                        "description": "Информация о последнем визите",
                        "type": "object",
                        "nullable": true
                    },
                    "followers_count": {
                        "description": "Количество подписчиков",
                        "type": "integer",
                        "example": 150,
                        "nullable": true
                    },
                    "city": {
                        "description": "Информация о городе",
                        "type": "object",
                        "nullable": true
                    },
                    "online": {
                        "description": "Статус онлайн",
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "birthday_date": {
                        "description": "Дата рождения",
                        "type": "string",
                        "example": "15.10.1990",
                        "nullable": true
                    },
                    "country": {
                        "description": "Информация о стране",
                        "type": "object",
                        "nullable": true
                    },
                    "sex": {
                        "description": "Пол пользователя (1 - женский, 2 - мужской)",
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "VkUserDataResponse": {
                "description": "Ответ с данными пользователя ВКонтакте",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BaseSuccessResponse"
                    },
                    {
                        "properties": {
                            "message": {
                                "description": "Сообщение о результате операции",
                                "type": "string",
                                "example": "Информация о профиле получена"
                            },
                            "data": {
                                "$ref": "#/components/schemas/VkUserData"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        }
    },
...
}